changed RWMutex back to Mutex as I removed the only reader mutex I had. Also added proper locking around V8 Context creation and disposing.
A note on V8 terminology:
A "context" is a javascript context. Each context has its own global object and its own set of global functions. Contexts can communicate eachother (given you give permission to do so) but dont execute in parallel.
Each context have access to the engine at turns (it supports preemption with a ContextSwitcher object that forces the task-switch, otherwise it runs one script entirely until it finish before starting the other).
There're two classes that need to be used in order to support this schema: Locker and Unlocker. Locker is reentrant (like a critical section), Unlocker is not (remove all locks) and the ContextSwitcher can overtake this locks.

Typically you need to use a Locker around any function that "touchs" the context, including of course executing any kind of script.

On the other hand, an "isolate" is a complete engine (with physical threads supporting it), you can run multiple isolates in parallel without any kind of lock contention BUT they cannot communicate with another isolates (of course... they're isolated ;) )

May be I could support both: having a default isolate that handles the process events (new thread, context switch, whatever else) and an isolate per-thread with a single context on it for the instrumentation. In that case I'll need some way of communicating information with other threads though.