Made a list of EnsureContextCallback callbacks on the context manager.
This fixed some crashes caused by creating a pintool internal thread from outside the main thread (or threads created from main), which is specifically forbidden in the API. Now we enqueue a callback request and trigger the context manager processing loop. Which in turn creates the thread that waits for the context to be created and execute the callback.
Also added a small change in EnsureContext such that a call with a fixed create=false gets inlined as a single TlsGetValue operation, suitable for analysis functions that need to be as fast as possible and, as ThreadStart already got executed, we can assume fairly safely that the context is already set for it. If some unforeseen havoc happen, we can add a CMP there to bail out if it's not set already.