support for mksnapshot to work as a pintool.
This is needed to create a snapshoot  of the context WITH the PIN functions added, otherwise the runtime and inlined functions don't work.
This code runs correctly and generate the files but they're broken for some reason and the snapshot doesn't work, it would be nice to have it working but it's not mandatory, I switched to v8_nosnapshot_pin in the meantime.