Function cache optimization.
Moved the AnalysisFunction cache to the PinContext class, it makes more sense to be there and it allows to have multiple AFs which source code (the body) is equal, so we just maintain one Function for multiple AFs. The example code shows this behaviour, one AF source but two functions with different arguments (but EnsureFunction is called just once).
Also improved the uncaught exception info shown.
Removed some debugging code and added a Hash function for the AF body text (Paul Larson hash).